Fix ternary operator auto-boxing bug

+++ src/main/java/net/snet/crm/service/utils/Entities.java
@@ -165,7 +165,7 @@ public class Entities {
 
 
     public Long asLongOr(final Long fallBack) {
-      return isNullOrEmpty() ? fallBack : asLong();
+      return isNullOrEmpty() ? fallBack : (Long) asLong();
     }
